Making sure all the settings are of the correct type

+++ core/src/main/java/dev/ratas/aggressiveanimals/aggressive/settings/type/MobAttackSettings.java
@@ -1,14 +1,14 @@
 package dev.ratas.aggressiveanimals.aggressive.settings.type;
 
 public record MobAttackSettings(Setting<Double> damage, Setting<Double> attackDamageLimit, Setting<Double> speed,
-        Setting<Float> attackLeapHeight) implements CheckingSettingBoundle {
+        Setting<Double> attackLeapHeight) implements CheckingSettingBoundle {
 
     @Override
     public void checkAllTypes() throws IllegalStateException {
         checkType(damage, Double.class);
         checkType(attackDamageLimit, Double.class);
         checkType(speed, Double.class);
-        checkType(attackLeapHeight, Float.class);
+        checkType(attackLeapHeight, Double.class);
     }
 
 }
